Add Quaternion files
Now the Test Artifact is not "Unit Test", instead is "Android Instrumentation Tests". Both tests `ApplicationTest.java` and `TestQuaternion.java` run on device or virtual device, but not in console as before. However, if no device is connected, the `@RequiresDevice` test is not run.

It would be nice to be able to run some tests in console as with the "Unit Test" Test Artifact, but here. Or to be able to switch between them in the same project.